+++ lib/webspicy/scope.rb
@@ -48,20 +48,20 @@ module Webspicy
 
     def each_example(service)
       service.examples.select(&to_filter_proc(config.test_case_filter)).each{|e|
-        yield(expand_example(service, e), false)
+        yield(expand_example(service, e))
       }
     end
 
     def each_counterexamples(service, &bl)
       service.counterexamples.select(&to_filter_proc(config.test_case_filter)).each{|e|
-        yield(expand_example(service, e), true)
+        yield(expand_example(service, e))
       } if config.run_counterexamples?
     end
 
     def each_generated_counterexamples(service, &bl)
       Webspicy.with_scope(self) do
         service.generated_counterexamples.select(&to_filter_proc(config.test_case_filter)).each{|e|
-          yield(expand_example(service, e), true)
+          yield(expand_example(service, e))
         }
       end if config.run_counterexamples?
     end